refactor(website): use browser test instead
